+++ KeeLocker/KeeLocker.csproj
@@ -57,6 +57,7 @@
       <DependentUpon>KeeLockerEntryTab.cs</DependentUpon>
     </Compile>
     <Compile Include="Fveapi.cs" />
+    <Compile Include="Common.cs" />
     <Compile Include="Plugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Forms\RichComboBox.cs">
